test(schema): golden-ify the base schema allowlist
Accepting an intentional base-view change previously meant hand-
editing a 440KB single-line JSON file. Assert the base view as SDL
against a golden under testdata/ instead, regenerated with -update
like the other golden tests; the failure message keeps the view-
gating guidance for new APIs and points at the regen command.

The SDL is rendered by the same printer that generates the docs'
schema.graphqls, extracted from cmd/introspect into an importable
cmd/codegen/introspection/sdl package; the introspect binary's output
is byte-identical before and after. SDL cuts the allowlist to about a
third of the old JSON's size, and future diffs read like schema
changes instead of introspection plumbing.
